Make the REST controller ask for authorization

+++ src/Rest/AbstractController.php
@@ -17,6 +17,19 @@ use Framewub\Http\Message\Response\Json as JsonResponse;
 
 abstract class AbstractController
 {
+    const FIND_ALL = 1;
+    const FIND_BY_ID = 2;
+    const ADD = 3;
+    const UPDATE = 4;
+    const DELETE = 5;
+
+    /**
+     * The status code of the response
+     *
+     * @var int
+     */
+    protected $statusCode = 200;
+
     /**
      * Handles a request and gives a response
      *
@@ -34,24 +47,44 @@ abstract class AbstractController
         switch ($request->getMethod()) {
             case 'GET':
                 if ($id) {
-                    $result = $this->findById($id);
+                    if ($this->isAuthorized(static::FIND_BY_ID)) {
+                        $result = $this->findById($id);
+                    } else {
+                        $this->statusCode = 403;
+                    }
                 } else {
-                    $result = $this->findAll();
+                    if ($this->isAuthorized(static::FIND_ALL)) {
+                        $result = $this->findAll();
+                    } else {
+                        $this->statusCode = 403;
+                    }
                 }
                 break;
 
             case 'PUT':
             case 'POST':
                 if ($id) {
-                    $result = $this->update($id, $request->getParsedBody());
+                    if ($this->isAuthorized(static::UPDATE)) {
+                        $result = $this->update($id, $request->getParsedBody());
+                    } else {
+                        $this->statusCode = 403;
+                    }
                 } else {
-                    $result = $this->add($request->getParsedBody());
+                    if ($this->isAuthorized(static::ADD)) {
+                        $result = $this->add($request->getParsedBody());
+                    } else {
+                        $this->statusCode = 403;
+                    }
                 }
                 break;
 
             case 'DELETE':
                 if ($id) {
-                    $result = $this->delete($id);
+                    if ($this->isAuthorized(static::DELETE)) {
+                        $result = $this->delete($id);
+                    } else {
+                        $this->statusCode = 403;
+                    }
                 }
                 break;
         }
@@ -60,7 +93,28 @@ abstract class AbstractController
             $result = $result->toArray();
         }
 
-        return new JsonResponse($result);
+        $response = new JsonResponse($result);
+        if ($this->statusCode != 200) {
+            $response = $response->withStatus($this->statusCode);
+        }
+
+        return $response;
+    }
+
+    /**
+     * Checks if the client is authorized to perform the specified action. Each
+     * overriding class should implement this in its own way, this method is
+     * just a stub which always returns true.
+     *
+     * @param int $action
+     *   One of the action constants of this class
+     *
+     * @return bool
+     *   Returns true if the client is authorized, false if not.
+     */
+    protected function isAuthorized($action)
+    {
+        return true;
     }
 
     /**
